// frontend/src/lib/generated/api/models/DTOAlumno.ts
export interface DTOAlumno {
    /**
     * Unique student ID
     * @type {number}
     * @memberof DTOAlumno
     */
    id: number;
    /**
     * Unique username
     * @type {string}
     * @memberof DTOAlumno
     */
    username: string;
    /**
     * Student's first name
     * @type {string}
     * @memberof DTOAlumno
     */
    firstName: string;
    /**
     * Student's last name
     * @type {string}
     * @memberof DTOAlumno
     */
    lastName: string;
    /**
     * Student's DNI
     * @type {string}
     * @memberof DTOAlumno
     */
    dni: string;
    /**
     * Student's email
     * @type {string}
     * @memberof DTOAlumno
     */
    email: string;
    /**
     * Phone number
     * @type {string}
     * @memberof DTOAlumno
     */
    phoneNumber?: string;
    /**
     * Enrollment date
     * @type {Date}
     * @memberof DTOAlumno
     */
    enrollmentDate?: Date;
    /**
     * Indicates if the student is enrolled
     * @type {boolean}
     * @memberof DTOAlumno
     */
    enrolled: boolean;
    /**
     * Indicates if the account is enabled
     * @type {boolean}
     * @memberof DTOAlumno
     */
    enabled: boolean;
    /**
     * List of enrolled class IDs
     * @type {Array<number>}
     * @memberof DTOAlumno
     */
    classIds?: Array<number>;
    /**
     * List of payment IDs
     * @type {Array<number>}
     * @memberof DTOAlumno
     */
    paymentIds?: Array<number>;
    /**
     * List of exercise submission IDs
     * @type {Array<number>}
     * @memberof DTOAlumno
     */
    submissionIds?: Array<number>;
    /**
     * User role
     * @type {string}
     * @memberof DTOAlumno
     */
    role: DTOAlumnoRoleEnum;
    /**
     * 
     * @type {string}
     * @memberof DTOAlumno
     */
    fullName?: string;
    /**
     * 
     * @type {number}
     * @memberof DTOAlumno
     */
    paymentCount?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOAlumno
     */
    classCount?: number;
    /**
     * 
     * @type {number}
     * @memberof DTOAlumno
     */
    submissionCount?: number;
}


/**
 * @export
 */
export const DTOAlumnoRoleEnum = {
    Admin: 'ADMIN',
    Profesor: 'PROFESOR',
    Alumno: 'ALUMNO',
    Usuario: 'USUARIO'
} as const;
export type DTOAlumnoRoleEnum = typeof DTOAlumnoRoleEnum[keyof typeof DTOAlumnoRoleEnum];